Немного поправил файловую структуру
