+++ src/utils/render-DOM-elements.js
@@ -1,0 +1,14 @@
+const render = (parent, element, position) => {
+
+  if (position === 'beforeend') {
+    parent.append(element);
+  } else {parent.prepend(element);}
+};
+
+const createElementDOM = (template) => {
+  const parent = document.createElement('div');
+  parent.innerHTML = template;
+  return parent.firstChild;
+};
+
+export {render, createElementDOM};
