+++ src/view/trip-price.js
@@ -1,5 +1,15 @@
-import {sumTripPrice} from '../utils.js';
 import AbstractView from './abstract-view.js';
+
+const sumTripPrice = (array) => {
+  return array.reduce((accumulator,obj) => {
+    const offersPriceSum =  obj.options.reduce((accumulator, option) => {
+      return accumulator + option.price;
+    }, 0);
+
+    return accumulator + obj.price + offersPriceSum;
+  }, 0);
+};
+
 
 const createTripPrice = (array) => {
   const price = sumTripPrice(array);
